+++ packages/vue-query/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @tanstack/vue-query
 
+## 5.99.2
+
+### Patch Changes
+
+- fix(vue-query): allow computed ref and other reactive values as `enabled` property in queryOptions ([#10465](https://github.com/TanStack/query/pull/10465))
+
+  This fixes a regression introduced in #10452 where `queryOptions` only accepted getter functions for the `enabled` property, but not `computed` refs or other reactive values.
+
+  Now the `enabled` property in `queryOptions` correctly accepts:
+  - `boolean` values
+  - `Ref<boolean>`
+  - `ComputedRef<boolean>`
+  - `() => boolean` getter functions
+  - `(query) => boolean` query predicate functions
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.99.2
+
+## 5.99.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.99.1
+
 ## 5.99.0
 
 ### Minor Changes